# [ready] perf: simpler Tensor init