@tinyb0t 2023-08-26
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/tensor.py                         7              15
------------------------------------------------------------
lines added in the tinygrad folder: -8
```